Merge the `chandra` VAMP feed into `studieschandra`
This brings the number of `trace` output lines down to 1470. This also
means that we won't be updating the `chandra` VAMP feed file with this
system, but it hasn't been getting updated anyway, and I'd rather stick
with the `studieschandra` file for consistency on Windows where it
probably matters more.