Rename two Saturn images and move one
These were just named "Saturn", causing problems for the `GotoSaturn`
command functionality as reported in
WorldWideTelescope/wwt-windows-client#210. I'm pretty sure that a simple
rename should fix the issue.

The astrometry for the "Infrared Ring Around Saturn" looks to be
ballpark OK -- the top half of the image seems consistent with PS1-3pi,
although with a substantial shift. "Big Band of Dust" doesn't agree
astrometrically and it is mostly non-sky data, so we archive it.