catfiles/explorerootweb.yml: add this
It's been true for a little while, but with this framework we can move
away from the customized webclient explore root, and move towards
centralized management of the corpus roots. I think it will still be
valuable for the webclient to have its own "Explore root" since it
doesn't have the back-compat restrictions of Windows, but we can at
least manage the files in a common framework.