studiesradio: add Places around some bare imagesets
The webclient doesn't properly display these images when you click on
them in the Explorer view. I think the policy is that an imageset not
contained in a Place is a "survey", which gets loaded as a background
image; and that makes sense to a certain extent. So let's manufacture
some simple Place data for these images, which should definitely not be
treated as surveys.