Remove duplicated Spitzer images
It turns out that the Spitzer VAMP dataset and the "classic" Spitzer
data contained a lot of duplicate images. Here we remove them.

It's a bit of judgment call as to whether to keep the VAMP or the
classic images. The classic images have nicer metadata, generally, and
it's possible that they have slightly better astrometric registration.
But I'm preferring the VAMP images because they live on our own servers,
whereas a lot of the classic images are "virtualearth.net" ones that I
worry might disappear at some point.