Start cleaning up Chandra catalogs
It looks like we have a few problems here. As with Spitzer, there are
duplicated images between the VAMP feeds and the original
virtualearth.net corpus.

Secondly, many of the VAMP feed items seem to be mislabeled with the
wrong "credits URL". This is pretty important since that's what
basically identifies the actual image.